+++ src-gui/src/renderer/components/pages/help/DaemonControlBox.tsx
@@ -2,7 +2,7 @@ import { Box, makeStyles } from "@material-ui/core";
 import FolderOpenIcon from "@material-ui/icons/FolderOpen";
 import PlayArrowIcon from "@material-ui/icons/PlayArrow";
 import PromiseInvokeButton from "renderer/components/PromiseInvokeButton";
-import { useAppSelector, useIsContextAvailable } from "store/hooks";
+import { useAppSelector } from "store/hooks";
 import InfoBox from "../../modal/swap/InfoBox";
 import CliLogsBox from "../../other/RenderedCliLog";
 import { initializeContext } from "renderer/rpc";
@@ -20,6 +20,8 @@ const useStyles = makeStyles((theme) => ({
 export default function DaemonControlBox() {
   const classes = useStyles();
   const logs = useAppSelector((s) => s.rpc.logs);
+
+  // The daemon can be manually started if it has failed or if it has not been started yet
   const canContextBeManuallyStarted = useAppSelector(
     (s) => s.rpc.status?.type === "Failed" || s.rpc.status === null,
   );
